server/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
Flask-JWT-Extended>=4.5.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
marshmallow==3.20.1
marshmallow-sqlalchemy==0.28.1
python-dotenv==1.0.0
sendgrid==6.11.0
cloudinary==2.7.0
Pillow==10.0.0
Flask-Cors==4.0.0
Werkzeug==3.0.0
